修改代码规范问题

+++ src/main/java/avoid_double_equals_for_objects/128/avoidDoubleEqualsForObjects128/src/main/java/org/example/Main.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 
 /**
  * 规范：对象比较必须使用 equals 而不是 "=="
- * 违规说明：在 if 条件中使用 "==" 比较 字节数组 (byte[]) 对象引用
- * 测试意图：覆盖比较对象类型：字节数组 (byte[])
+ * 违规说明：在 if 条件中使用 "==" 比较 短整型数组 (short[]) 对象引用
+ * 测试意图：覆盖比较对象类型：短整型数组 (short[])
  *
  * @author 曹卓熠
  * @version 1.0.0
@@ -20,8 +20,8 @@ public class Main {
      * @param args 命令行参数
      */
     public static void main(String[] args) {
-        byte[] arr1 = new byte[1]; // 测试因子(compared_object_type=array_primitive_byte)
-        byte[] arr2 = new byte[1];
+        short[] arr1 = new short[1]; // 测试因子(compared_object_type=array_primitive_short)
+        short[] arr2 = new short[1];
         if (arr1 == arr2) { // 检查点
             log.info("viol");
         }
