+++ src/main/java/avoid_double_equals_for_objects/130/avoidDoubleEqualsForObjects130/src/main/java/org/example/Main.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 
 /**
  * 规范：对象比较必须使用 equals 而不是 "=="
- * 违规说明：在 if 条件中使用 "==" 比较 整型数组 (int[]) 对象引用
- * 测试意图：覆盖比较对象类型：整型数组 (int[])
+ * 违规说明：在 if 条件中使用 "==" 比较 长整型数组 (long[]) 对象引用
+ * 测试意图：覆盖比较对象类型：长整型数组 (long[])
  *
  * @author 曹卓熠
  * @version 1.0.0
@@ -20,8 +20,8 @@ public class Main {
      * @param args 命令行参数
      */
     public static void main(String[] args) {
-        int[] arr1 = new int[1]; // 测试因子(compared_object_type=array_primitive_int)
-        int[] arr2 = new int[1];
+        long[] arr1 = new long[1]; // 测试因子(compared_object_type=array_primitive_long)
+        long[] arr2 = new long[1];
         if (arr1 == arr2) { // 检查点
             log.info("viol");
         }
